Add webpack, form functionality
In this patch, I went and got the modules talking to each other with
webpack. In the DOM file, we just added some functionality to push the
new To Do item into an array, and we have a button that lists each To Do
item right now. Next step is updating the DOM to include the contents
of the items, rather than just console logging them.